Move EntityPermissionChecker from framework/entityext to application/content (OFBIZ-13393) Relocated EntityPermissionChecker from framework/entityext to application/content due to its strong dependency on the Content data model. - Moved EntityPermissionChecker implementation to applications/content - Removed framework-level dependencies related to EntityPermissionChecker - Removed if-entity-permission handling from framework code - Reduced framework coupling with Content application entities - Improved separation between framework and application-specific functionality
