+++ framework/widget/src/main/java/org/apache/ofbiz/widget/model/AbstractModelCondition.java
@@ -38,7 +38,6 @@ import org.apache.ofbiz.base.util.UtilXml;
 import org.apache.ofbiz.base.util.collections.FlexibleMapAccessor;
 import org.apache.ofbiz.base.util.string.FlexibleStringExpander;
 import org.apache.ofbiz.entity.GenericValue;
-import org.apache.ofbiz.entityext.permission.EntityPermissionChecker;
 import org.apache.ofbiz.minilang.operation.BaseCompare;
 import org.apache.ofbiz.security.Security;
 import org.apache.ofbiz.service.DispatchContext;
@@ -209,8 +208,6 @@ public abstract class AbstractModelCondition implements Serializable, ModelCondi
                 return new IfRegexp(factory, modelWidget, conditionElement);
             } else if ("if-empty".equals(nodeName)) {
                 return new IfEmpty(factory, modelWidget, conditionElement);
-            } else if ("if-entity-permission".equals(nodeName)) {
-                return new IfEntityPermission(factory, modelWidget, conditionElement);
             } else {
                 throw new IllegalArgumentException("Condition element not supported with name: " + conditionElement.getNodeName());
             }
@@ -404,33 +401,6 @@ public abstract class AbstractModelCondition implements Serializable, ModelCondi
             return fieldAcsr;
         }
 
-    }
-
-    /**
-     * Models the &lt;if-entity-permission&gt; element.
-     * @see <code>widget-common.xsd</code>
-     */
-    public static final class IfEntityPermission extends AbstractModelCondition {
-        private final EntityPermissionChecker permissionChecker;
-
-        private IfEntityPermission(ModelConditionFactory factory, ModelWidget modelWidget, Element condElement) {
-            super(factory, modelWidget, condElement);
-            this.permissionChecker = new EntityPermissionChecker(condElement);
-        }
-
-        @Override
-        public void accept(ModelConditionVisitor visitor) throws Exception {
-            visitor.visit(this);
-        }
-
-        @Override
-        public boolean eval(Map<String, Object> context) {
-            return permissionChecker.runPermissionCheck(context);
-        }
-
-        public EntityPermissionChecker getPermissionChecker() {
-            return permissionChecker;
-        }
     }
 
     /**
